refactor(legacy support): use unified component
* uses `in-element-polyfill` and `maybe-in-element` AST transforms to allow to use `in-element` with legacy support back to Ember 1.13 (ember-wormhole based)
* Cleaned up unnecessary legacy components

Closes #71

BREAKING CHANGE:
Support for Ember <1.13 has been dropped!
