--- addon/templates/components/-ember-popper-legacy-1.11.hbs
@@ -1,1 +1,0 @@
-{{yield _popperHash}}
