--- addon/templates/components/-ember-popper-legacy.hbs
@@ -1,6 +1,0 @@
-{{yield (hash
-  disableEventListeners=(action 'disableEventListeners')
-  enableEventListeners=(action 'enableEventListeners')
-  scheduleUpdate=(action 'scheduleUpdate')
-  update=(action 'update')
-)}}
