+++ addon/templates/components/ember-popper.hbs
@@ -1,30 +1,17 @@
 {{unbound _parentFinder}}
 
-{{#if renderInPlace}}
+{{!-- Elements that exist deep within the document tree are given an implicitly lower z-index
+      value than elements that aren't as deep in the tree; this has the effect of hiding our
+      ember-popper behind less-nested elements. Due to the way z-indexing works, we cannot simply
+      add a higher z-index value to our ember-popper. To avoid this issue, we render the element
+      on the document body, giving it the highest default z-index value. --}}
+{{#maybe-in-element _popperContainer renderInPlace}}
   <div id={{id}} class={{class}} role={{ariaRole}}>
     {{yield (hash
-      disableEventListeners=(action 'disableEventListeners')
-      enableEventListeners=(action 'enableEventListeners')
-      scheduleUpdate=(action 'scheduleUpdate')
-      update=(action 'update')
-    )}}
+              disableEventListeners=(action 'disableEventListeners')
+              enableEventListeners=(action 'enableEventListeners')
+              scheduleUpdate=(action 'scheduleUpdate')
+              update=(action 'update')
+            )}}
   </div>
-{{else}}
-  {{!-- Elements that exist deep within the document tree are given an implicitly lower z-index
-        value than elements that aren't as deep in the tree; this has the effect of hiding our
-        ember-popper behind less-nested elements. Due to the way z-indexing works, we cannot simply
-        add a higher z-index value to our ember-popper. To avoid this issue, we render the element
-        on the document body, giving it the highest default z-index value. --}}
-
-  {{#-in-element _popperContainer}}
-    {{!-- Add a wrapper around the yielded block so we have something for the Popper to target --}}
-    <div id={{id}} class={{class}} role={{ariaRole}}>
-      {{yield (hash
-        disableEventListeners=(action 'disableEventListeners')
-        enableEventListeners=(action 'enableEventListeners')
-        scheduleUpdate=(action 'scheduleUpdate')
-        update=(action 'update')
-      )}}
-    </div>
-  {{/-in-element}}
-{{/if}}
+{{/maybe-in-element}}
