+++ package.json
@@ -31,10 +31,11 @@
     "ember-cli-babel": "^6.10.0",
     "ember-cli-htmlbars": "^2.0.3",
     "ember-cli-node-assets": "^0.2.2",
-    "ember-cli-version-checker": "^2.1.0",
-    "ember-compatibility-helpers": "^0.1.3",
     "ember-decorators": "^1.3.2",
+    "ember-hash-helper-polyfill": "^0.2.1",
+    "ember-in-element-polyfill": "^0.1.2",
     "ember-legacy-class-transform": "^0.1.4",
+    "ember-maybe-in-element": "^0.1.3",
     "ember-raf-scheduler": "^0.1.0",
     "fastboot-transform": "^0.1.0",
     "popper.js": "^1.12.9"
@@ -53,7 +54,6 @@
     "ember-cli-uglify": "^2.0.0",
     "ember-disable-prototype-extensions": "^1.1.3",
     "ember-export-application-global": "^2.0.0",
-    "ember-hash-helper-polyfill": "^0.2.0",
     "ember-load-initializers": "^1.0.0",
     "ember-native-dom-event-dispatcher": "^0.6.3",
     "ember-native-dom-helpers": "^0.5.8",
