+++ testem.js
@@ -15,7 +15,8 @@ module.exports = {
         '--disable-gpu',
         '--headless',
         '--remote-debugging-port=0',
-        '--window-size=1440,900'
+        '--window-size=1440,900',
+        '--no-sandbox'
       ]
     }
   }
